+++ src/controllers/transaction/show-transaction.controller.js
@@ -1,0 +1,70 @@
+require("dotenv").config();
+const axios = require("axios");
+const { PrismaClient } = require("@prisma/client");
+const prisma = new PrismaClient();
+const jwt = require("jsonwebtoken");
+const DateTimeUtils = require("../../libs/datetime");
+const Currency = require("../../libs/currency");
+const Moment = require("../../libs/moment");
+
+const { STAGING_SERVER_NO_SSL, PRODUCTION_SERVER_NO_SSL, NODE_ENV } =
+  process.env;
+
+const MIDTRANS_API_URL =
+  "https://app.sandbox.midtrans.com/snap/v1/transactions";
+const MIDTRANS_SERVER_KEY = process.env.MIDTRANS_SERVER_KEY;
+function calculateTotalPrice(booking) {
+  return Math.floor(booking.booking_amount);
+}
+
+class ShowTransactionController {
+  static async showTransaksi(req, res) {
+    try {
+      const transaksi = await prisma.bookings.findMany();
+      if (!transaksi || transaksi.length === 0) {
+        return res.status(404).json({
+          statusCode: 404,
+          status: "failed",
+          message: "Tidak ada transaksi ditemukan",
+          data: [],
+        });
+      }
+
+      res.status(200).json({
+        statusCode: 200,
+        status: "success",
+        message: "Berhasil menampilkan list transaksi",
+        data: transaksi,
+      });
+    } catch (error) {
+      console.error("Error fetching transactions:", error);
+
+      // Menangani kesalahan koneksi database
+      if (error instanceof Prisma.PrismaClientKnownRequestError) {
+        return res.status(400).json({
+          statusCode: 400,
+          status: "failed",
+          message: "Kesalahan dalam permintaan ke database",
+          details: error.message,
+        });
+      }
+      // Menangani kesalahan koneksi database
+      if (error instanceof Prisma.PrismaClientInitializationError) {
+        return res.status(503).json({
+          statusCode: 503,
+          status: "failed",
+          message: "Kesalahan koneksi ke database",
+          details: error.message,
+        });
+      }
+      res.status(500).json({
+        statusCode: 500,
+        status: "failed",
+        message: "Terjadi kesalahan pada server",
+        details: error.message,
+      });
+    }
+  }
+}
+
+module.exports = new ShowTransactionController();
